spatstat.sparse-internal {spatstat.sparse}	R Documentation

Internal Functions of spatstat.sparse Package

Description

Internal utility functions of the spatstat.sparse package.

Usage

representativeRows(x)

checksolve(M, action, descrip, target)
check.mat.mul(A, B, Acols, Brows, fatal)
sumsymouter(x, w, distinct)

unionOfSparseIndices(A,B)
inside3Darray(d, i, j, k)
SparseEntries(x)
SparseIndices(x)
EntriesToSparse(df, dims)
mapSparseEntries(x, margin, values, conform, across)
applySparseEntries(x, f, ...)
sumsymouterSparse(x, w, distinct, dbg)
rbindCompatibleDataFrames(x)
check.anySparseVector(v, npoints, fatal, things,
                      naok, warn, vname, oneok)
evalSparse3Dentrywise(expr, envir)
expandSparse(x, n, across)
sparseVectorCumul(x, i, length)
tensor1x1(A, B)

Details

These internal spatstat.sparse functions are not usually called directly by the user. Their names and capabilities may change without warning from one version of spatstat.utils to the next.

[Package spatstat.sparse version 3.1-0 Index]